Move sources model to dbt package
