+++ internal/pkg/dbt/dbt.go
@@ -7,6 +7,7 @@ import (
 )
 
 const ProjectFile = "dbt_project.yml"
+const SourcesPath = "models/_sources"
 
 type Project struct {
 	fs filesystem.Fs
@@ -16,7 +17,7 @@ func LoadProject(fs filesystem.Fs) (*Project, error) {
 	if !fs.Exists(ProjectFile) {
 		return nil, fmt.Errorf(`missing file "%s" in the "%s"`, ProjectFile, fs.BasePath())
 	}
-	
+
 	// TODO
 
 	return &Project{fs: fs}, nil
